Cleaned up functional tests and started adding LiveServerTestCase
Moved functional tests for both lists and polls sites into a \mysite_proj\
functional_test directory.  Began working on LiveServerTestCase method
to run functional tests, which will create a temporary data set for testing
that is discarded after the test run.
