--- polls/functional_tests.py
@@ -1,50 +1,0 @@
-__author__ = 'cbinder'
-
-from selenium import webdriver
-import unittest
-
-class NewVisitorTest(unittest.TestCase):
-    def setUp(self):  #2
-        #chrome_option = webdriver.ChromeOptions()
-        #chrome_option.add_argument('--proxy-server=us-auto.proxy.lexmark.com:80' )
-        self.browser = webdriver.Chrome()
-        self.browser.implicitly_wait(3)
-        #self.browser.service()
-        #self.browser.
-
-    def tearDown(self):  #3
-        self.browser.close()
-        #self.browser.quit()
-
-    def test_can_start_a_list_and_retrieve_it_later(self):  #4
-        # Edith has heard about a cool new online to-do app. She goes
-        # to check out its homepage
-        self.browser.get('http://localhost:8000/polls')
-        #self.browser.get('http://www.google.com')
-
-        # She notices the page title and header mention to-do lists
-        self.assertIn('YO YO YO', self.browser.title)  #5
-        #self.fail('Finish the test!')  #6
-
-        # She is invited to enter a to-do item straight away
-
-        # She types "Buy peacock feathers" into a text box (Edith's hobby
-        # is tying fly-fishing lures)
-
-        # When she hits enter, the page updates, and now the page lists
-        # "1: Buy peacock feathers" as an item in a to-do list
-
-        # There is still a text box inviting her to add another item. She
-        # enters "Use peacock feathers to make a fly" (Edith is very methodical)
-
-        # The page updates again, and now shows both items on her list
-
-        # Edith wonders whether the site will remember her list. Then she sees
-        # that the site has generated a unique URL for her -- there is some
-        # explanatory text to that effect.
-
-        # She visits that URL - her to-do list is still there.
-
-        # Satisfied, she goes back to sleep
-
-
